backend/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
torch
transformers
numpy
opencv-python-headless
